refactor: MenuView에 delegate들 분리
